Update editing rights for a page
The name and url attributes of system pages can no longer be edited by a user
that does not have the SuperAdmin role. This means that url aliases for a
page that have been defined by the SuperAdmin will allways keep on
working. They also dont have to be defined as a reserved url.